refactor: move mappers to prisma folder
