+++ src/infrastructure/persistence/prisma/mappers/prisma-clinician-mapper.ts
