+++ src/infrastructure/persistence/prisma/mappers/prisma-medical-record-mapper.ts
